Support for boolean descriptors
