+++ examples_old/example4.py
@@ -1,5 +1,5 @@
-__author__ = 'github.com/wardsimon'
-__version__ = '0.1.0'
+__author__ = "github.com/wardsimon"
+__version__ = "0.1.0"
 
 import json
 from typing import Callable, List
@@ -23,6 +23,7 @@ class Calculator1:
     """
     Generic calculator in the style of crysPy
     """
+
     def __init__(self, m: float = 1, c: float = 0):
         """
         Create a calculator object with m and c
@@ -52,11 +53,10 @@ class Calculator2:
     Isolated calculator. This calculator can't have values set, it can
     only load/save data and calculate from it. i.e in the style of crysfml
     """
+
     def __init__(self):
-        """
-        """
-        self._data = {'m': 0,
-                      'c': 0}
+        """ """
+        self._data = {"m": 0, "c": 0}
 
     def calculate(self, x_array: np.ndarray) -> np.ndarray:
         """
@@ -67,7 +67,7 @@ class Calculator2:
         :return: points calculated at `x`
         :rtype: np.ndarray
         """
-        return self._data['m'] * x_array + self._data['c']
+        return self._data["m"] * x_array + self._data["c"]
 
     def export_data(self) -> str:
         return json.dumps(self._data)
@@ -80,6 +80,7 @@ class InterfaceTemplate(MSONable, metaclass=ABCMeta):
     """
     This class is a template and defines all properties that an interface should have.
     """
+
     _interfaces = []
     _borg = borg
 
@@ -141,6 +142,7 @@ class Interface1(InterfaceTemplate):
     """
     A simple example interface using Calculator1
     """
+
     def __init__(self):
         # This interface will use calculator1
         self.calculator = Calculator1()
@@ -168,7 +170,7 @@ class Interface1(InterfaceTemplate):
         :rtype: noneType
         """
         if self._borg.debug:
-            print(f'Interface1: Value of {value_label} set to {value}')
+            print(f"Interface1: Value of {value_label} set to {value}")
         setattr(self.calculator, value_label, value)
 
     def fit_func(self, x_array: np.ndarray) -> np.ndarray:
@@ -189,6 +191,7 @@ class Interface2(InterfaceTemplate):
     transfer it to the calculator (get and calculate) and import data
     from the calculator (set)
     """
+
     def __init__(self):
         """
         Set up a calculator and a local dict
@@ -220,7 +223,7 @@ class Interface2(InterfaceTemplate):
         :rtype: noneType
         """
         if self._borg.debug:
-            print(f'Interface2: Value of {value_label} set to {value}')
+            print(f"Interface2: Value of {value_label} set to {value}")
         self._data = json.loads(self.calculator.export_data())
         if value_label in self._data.keys():
             self._data[value_label] = value
@@ -281,7 +284,9 @@ class InterfaceFactory:
         serialized = self.__interface_obj.as_dict()
         interfaces = self.available_interfaces
         if new_interface in interfaces:
-            self._current_interface: InterfaceTemplate = self._interfaces[interfaces.index(new_interface)]
+            self._current_interface: InterfaceTemplate = self._interfaces[
+                interfaces.index(new_interface)
+            ]
             self.__interface_obj = self._current_interface.from_dict(serialized)
         else:
             raise AttributeError
@@ -319,7 +324,7 @@ class InterfaceFactory:
         :return: points calculated at positional values `x`
         :rtype: np.ndarray
         """
-        return self.__interface_obj.fit_func(x_array, *args, **kwargs)
+        return self.__interface_obj.fit_function(x_array, *args, **kwargs)
 
     def generate_bindings(self, name):
         """
@@ -345,8 +350,10 @@ class InterfaceFactory:
         :return: function to get key
         :rtype: Callable
         """
+
         def inner(obj):
             obj().get_value(key)
+
         return lambda obj: inner(obj)
 
     @staticmethod
@@ -361,8 +368,10 @@ class InterfaceFactory:
         :return: function to set key
         :rtype: Callable
         """
+
         def inner(value):
             obj().set_value(key, value)
+
         return inner
 
 
@@ -370,8 +379,8 @@ class Line(BaseObj):
     """
     Simple descriptor of a line.
     """
-    _defaults = [Parameter('m', 1),
-                 Parameter('c', 0)]
+
+    _defaults = [Parameter("m", 1), Parameter("c", 0)]
 
     def __init__(self, interface_factory: InterfaceFactory = None):
         """
@@ -380,8 +389,7 @@ class Line(BaseObj):
         :param interface_factory: interface controller object
         :type interface_factory: InterfaceFactory
         """
-        super().__init__(self.__class__.__name__,
-                         *self._defaults)
+        super().__init__(self.__class__.__name__, *self._defaults)
         self.interface = interface_factory
         self._set_interface()
 
@@ -395,19 +403,19 @@ class Line(BaseObj):
     @property
     def gradient(self):
         if self.interface:
-            return self.interface().get_value('m')
+            return self.interface().get_value("m")
         else:
             return self.m.raw_value
 
     @property
     def intercept(self):
         if self.interface:
-            return self.interface().get_value('c')
+            return self.interface().get_value("c")
         else:
             return self.c.raw_value
 
     def __repr__(self):
-        return f'Line: m={self.m}, c={self.c}'
+        return f"Line: m={self.m}, c={self.c}"
 
 
 interface = InterfaceFactory()
@@ -420,7 +428,7 @@ y = np.array([2, 4, 6], dtype=np.float64) - 1
 
 f_res = f.fit(x, y)
 
-print('\n######### Interface 1 #########\n')
+print("\n######### Interface 1 #########\n")
 print(f_res)
 print(line)
 
@@ -433,6 +441,6 @@ line.m.value = 1
 line.c.value = 0
 f_res = f.fit(x, y)
 
-print('\n######### Interface 2 #########\n')
+print("\n######### Interface 2 #########\n")
 print(f_res)
 print(line)
